iwlwifi: RX signal improvements.
Sum up multi-chain signals instead of just taking the maximum value.
This gives more correct total on-air signal value.

And, the firmware managed averages do not match expected per-packet values,
so ignore those for MQ drivers and instead calculate averages in the driver
and report that.

Combined with the previous patch, this gives better looking RX signal report:

	signal:  	-39 [-42, -41] dBm
	signal avg:	-39 [-42, -41] dBm
	beacon signal avg:	-46 dBm

iwlwifi: Disable beacon filtering by default.

So we can calculate the avg rx signal in the driver instead
of depending on the firmware.

iwlwifi: Fix per-chain rssi averaging for 1nss frames.

Averaging in the weak secondary chain for a single-spatial-stream packet
causes the per-chain averages to look abnormally weak.  So, move calculation
of nss above where the rssi is calculated, and then use the nss inside the
rssi-calculation to ignore the weak spatial stream for 1nss frames.

Signed-off-by: Ben Greear <greearb@candelatech.com>