wifi: mt76: mt7915: Support vht mu-mimo sniffer feature.
This feature allows mac80211 to update the driver with mu-mimo
group to allow the monitor port to capture MU-MIMO (VHT) frames.

The mt7915 driver implementation will only enable this feature
when there is only a monitor vdev.  I was afraid that it would mess
up a sta + monitor vdev combination, for instance.

Original code from Ryder

Signed-off-by: Ben Greear <greearb@candelatech.com>